requirements.txt
python-telegram-bot==21.6
sqlalchemy
requests
beautifulsoup4
psycopg2-binary
python-dotenv
pdfplumber
psutil